refactor: 치안 관련 오버레이 분리
